backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.12.1
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
httpx>=0.25.0
psycopg[binary,pool]>=3.1.12
openai>=0.28.1
python-dotenv>=1.0.0
apscheduler>=3.10.4
redis>=5.0.1
pytest>=7.4.3
pytest-asyncio>=0.21.1
black>=23.10.1
isort>=5.12.0
mypy>=1.6.1
